major cosmetics (interface change) + bugfix of 0.8.3
Omd_parser.parse in Omd_parser  have been moved to a  functor in order
to expose main_loop (and main_loop_rev) and to keep reentrance for the
users of the library.

In  order  not  to  impose  the  use  of  the  functor,  the  function
Omd_parser.default_parse   is    provided   as   a    replacement   of
Omd_parser.parse,  which  no longer  exists.  This renaming is  mainly
motivated  by the  fact that  there will  eventually be  many  ways to
instanciate the  parser (notably with extensions  and link definitions
environment) and the name should reflect this. ("will eventually be"
because the extension mechanism is not very mature yet)

The other files (omd.ml{,i} and  omd_main.ml) have been adapted to use
the new interface of Omd_parser.