Quick fix for -toc
A very quick-and-dirty fix for the -toc issue.
The intial problem is that -toc had actually no effect.
 (https://github.com/pw374/omd/issues/88)
The big nasty problem is that it seems that
Omd_parser.Make.make_paragraphs somehow clashes
with what's done for -toc (Omd_main.tag_toc?).
So it still needs to be properly fixed.