fix: link def environment for markdown in media:type="text/omd" sections
The link definition set was used too early when converting inner markdown
to HTML, so links defined later in the document were not available.
This fix delays the use of that environment so that when it's used,
it has all the link definitions.

This fixes https://github.com/pw374/omd/issues/102.