fix: make_paragraphs: distinct block vs. non-block for X(x) elements
Concerns Omd_parser.make_paragraphs:
If the extension returns a block as first element
(i.e., x#to_t returns (Some(a_block_element::_))),
then consider the extension as a block. However
don't take its contents as it is yet, the contents
of the extension shall be considered final as late
as possible.