rewrite: parser of inline HTML
The code that parses the inline HTML has been completely rewritten
from scratch. This uses the new representation of inline HTML chunks.
Wrongly balanced inline HTML is no longer seen as HTML at all.

This change needs more testing, as it has only been tested a little
(but so far it seems not too bad).
This is a big step towards version 1.0.0.