Use dune 2.7
This lets us use dune's built in cram testing, instead of the bespoke
cram-style testing that used for the spec. I think it makes sense to
support the current finicky testing setup for the spec, because it lets
us use the commonmark spec as a raw source of truth, but for further
tests we add I think we'll benefit from a more lightweight set up.

Since we are already using dune, it seems sensible to use this testing
framework we get access to with a two line change.

We might consider reworking the spec testing setup to generate
dune-supported cram tests in the future tool, as it would eliminate a
good chunk of code.