Rework element constructors
The constructors lacked some ergonomic convenience, and we were missing
a constructor for headings and missing a way to construct elements that
included concatenated inlines.

I think this was mostly due to us omitting the addition of tests that
would show how the constructors were meant to be used, so I'm adding in
tests as I go.